Fix MultiStream latency param to handle sub-millisec

+++ loadgen/test_settings.h
@@ -126,7 +126,7 @@ struct TestSettings {
   /// \brief A hint used by the loadgen to pre-generate enough samples to
   ///        meet the minimum test duration.
   /// \brief MultiStream latency is for query (not sample) latency
-  uint64_t multi_stream_expected_latency_ns = 8000000;
+  double multi_stream_expected_latency_ns = 8000000;
   /// \brief The latency percentile for MultiStream mode.
   double multi_stream_target_latency_percentile = 0.99;
   /// \brief The number of samples in each query.
